+++ src/main.rs
@@ -5,23 +5,22 @@
 //!
 //! Author: Sergei Abramenkov
 //! License: MIT
-//! Version: 0.0.3
+//! Version: 0.0.4
 //! ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 use std::env;
-use std::fs::{File, OpenOptions};
+use std::fs::OpenOptions;
 use std::io::{Read, Write};
 use std::path::PathBuf;
 
 const ARGUMENTS: &str = "
 Required arguments (must be named - not positional):
-  -d, --physical-drive <DRIVE>      Device identifier
-  -o, --output-path <PATH>          Output image file";
+  -s, --source <TARGET>             What to image
+  -d, --destination <TARGET>        Where to put";
 
 const FLAGS: &str = "
 Flags (boolean toggles - false by default):
   -h, --help                        Show extended help with examples
-  -f, --force                       Force overwrite output if exists
   -t, --smart                       SmarT Truncation (after *empty chunk)
                                       *consecutive 0x00 or 0xFF pattern";
 
@@ -40,33 +39,129 @@ enum HelpLevel {
     Extended,
 }
 
+impl HelpLevel {
+    fn display(&self) {
+        println!("rusdd - Really Useful Secure Digital Duplicator");
+        println!("\nUsage: rusdd [FLAGS] -d <DRIVE> -o <PATH> [PARAMS]");
+        match self {
+            HelpLevel::Usage => {
+                println!("{}", ARGUMENTS);
+                println!("{}", FLAGS);
+                println!("Try 'rusdd -h' for more information.");
+            }
+            HelpLevel::Extended => {
+                println!("{}", ARGUMENTS);
+                println!("{}", FLAGS);
+                println!("{}", PARAMS);
+            }
+        }
+    }
+}
+
+/// A higher level abstraction useful for better versatility
+/// "Imagine" this tool could be eventually used for both:
+///   (A) imaging some SD-card from a datalogger to HDD backup as a file
+///   (B) putting backup image "back" on the unit with a fresh SD-card
+/// Initially I was thinking (and it is still the core goal) of only A
+/// But if planned and architected smart - B could be feasible as well
+enum Location {
+    Drive(String),  // Actual drive: "\\.\PHYSICALDRIVE3" or /dev/sdc
+    Image(PathBuf), // Image file: "sd-card.dd"
+}
+
+impl Location {
+    fn from_str(way: String) -> Location {
+        #[cfg(target_os = "windows")]
+        {
+            if way.starts_with("\\\\.\\") {
+                return Location::Drive(way);
+            }
+        }
+
+        #[cfg(target_family = "unix")]
+        {
+            if way.starts_with("/dev/") {
+                return Location::Drive(way);
+            }
+        }
+
+        Location::Image(PathBuf::from(way))
+    }
+
+    fn open_for_read(&self) -> Result<impl Read, String> {
+        match self {
+            Location::Drive(way) => OpenOptions::new()
+                .read(true)
+                .open(way)
+                .map_err(|e| format!("Open drive {} for read: {}", way, e)),
+            Location::Image(way) => OpenOptions::new()
+                .read(true)
+                .open(way)
+                .map_err(|e| format!("Open {} for read: {}", way.display(), e)),
+        }
+    }
+
+    fn open_for_write(&self) -> Result<impl Write, String> {
+        match self {
+            Location::Drive(way) => OpenOptions::new()
+                .write(true)
+                .open(way)
+                .map_err(|e| format!("Open drive {} for write: {}", way, e)),
+            Location::Image(way) => {
+                if way.exists() {
+                    return Err(format!(
+                        "File ({}) already exist. Check and move (delete) it.",
+                        way.display()
+                    ));
+                }
+                OpenOptions::new()
+                    .write(true)
+                    .create(true)
+                    .truncate(true)
+                    .open(way)
+                    .map_err(|e| format!("Cannot create image file: {}", e))
+            }
+        }
+    }
+
+    fn display(&self) -> String {
+        match self {
+            Location::Drive(way) => format!("{} (drive)", way),
+            Location::Image(way) => format!("{} (file)", way.display()),
+        }
+    }
+}
+
 /// Command line interface configuration
 struct Config {
-    physical_drive: String,
-    output_path: PathBuf,
-    force: bool,      // Force overwrite of the output path?
+    source: Location,
+    destination: Location,
     smart: bool,      // Is smart stop trancation enabled?
     sector_size: u16, // Media physical sector size in bytes
     chunk_size: u16,  // Pattern-seeking chunk size in sectors
     buffer_size: u16, // Reading buffer size in chunks
 }
 
-/// Print help in STDERR so it would not pollute STDOUT
-fn print_help(level: HelpLevel) {
-    eprintln!("rusdd - Really Useful Secure Digital Duplicator");
-    eprintln!("\nUsage: rusdd [FLAGS] -d <DRIVE> -o <PATH> [PARAMS]");
+impl Config {
+    fn calc_sizes(&self) -> (usize, usize, usize) {
+        let sector: usize = self.sector_size as usize;
+        let chunk: usize = (self.chunk_size as usize) * sector;
+        let buffer: usize = (self.buffer_size as usize) * chunk;
+        (sector, chunk, buffer)
+    }
 
-    match level {
-        HelpLevel::Usage => {
-            eprintln!("{}", ARGUMENTS);
-            eprintln!("{}", FLAGS);
-            eprintln!("Try 'rusdd -h' for more information.");
-        }
-        HelpLevel::Extended => {
-            eprintln!("{}", ARGUMENTS);
-            eprintln!("{}", FLAGS);
-            eprintln!("{}", PARAMS);
-        }
+    // Program welcome message (a header in case of STDOUT redirection)
+    fn display(&self) {
+        let (sector, chunk, buffer) = self.calc_sizes();
+        println!("^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^");
+        println!("rusdd - Really Useful Secure Digital Duplicator (ver.0.0.4)");
+        println!("Source:       {}", self.source.display());
+        println!("Destination:  {}", self.destination.display());
+        println!("Smart mode:   {}", if self.smart { "ON" } else { "OFF" });
+        println!("Sector size: {} bytes", sector);
+        println!("Chunk size: {} KiB", chunk / 1024);
+        println!("Buffer size: {} MiB", buffer / 1024 / 1024);
+        println!("***********************************************************");
     }
 }
 
@@ -85,10 +180,9 @@ fn parse_cli_from_iter<I>(mut cli: I) -> Result<Config, String>
 where
     I: Iterator<Item = String>,
 {
-    let mut physical_drive = None;
-    let mut output_path = None;
+    let mut source = None;
+    let mut destination = None;
     // Assigning defaults for flags and params
-    let mut force = false; // Better safe than sorry (warn user)
     let mut smart = false; // Full drive image (forensic frendly)
     let mut sector_size: u16 = 512; // Typical physical sector size
     let mut chunk_size: u16 = 16384; // => 8 MiB (8_589_934_592 bytes)
@@ -96,15 +190,14 @@ where
 
     while let Some(arg) = cli.next() {
         match arg.as_str() {
-            "--physical-drive" | "-d" => {
-                let value = cli.next().ok_or("Missing --physical-drive value")?;
-                physical_drive = Some(value);
+            "--source" | "-s" => {
+                let value = cli.next().ok_or("Missing --source value")?;
+                source = Some(Location::from_str(value));
             }
-            "--output-path" | "-o" => {
-                let value = cli.next().ok_or("Missing --output-path value")?;
-                output_path = Some(PathBuf::from(value));
+            "--destination" | "-d" => {
+                let value = cli.next().ok_or("Missing --destination value")?;
+                destination = Some(Location::from_str(value));
             }
-            "--force" | "-f" => force = true,
             "--smart" | "-t" => smart = true,
             "--sector-size" => {
                 let value = cli.next().ok_or("Missing --sector-size value")?;
@@ -119,7 +212,7 @@ where
                 buffer_size = parse_size_param("buffer size", value)?;
             }
             "--help" | "-h" => {
-                print_help(HelpLevel::Extended);
+                HelpLevel::Extended.display();
                 std::process::exit(0);
             }
             _ => return Err(format!("Unknown argument: {}", arg)),
@@ -127,9 +220,8 @@ where
     }
 
     Ok(Config {
-        physical_drive: physical_drive.ok_or("--physical-drive is required")?,
-        output_path: output_path.ok_or("--output-path is required")?,
-        force,
+        source: source.ok_or("--source is required")?,
+        destination: destination.ok_or("--destination is required")?,
         smart,
         sector_size: sector_size,
         chunk_size: chunk_size,
@@ -139,7 +231,7 @@ where
 
 fn parse_cli() -> Result<Config, String> {
     if env::args().len() == 1 {
-        print_help(HelpLevel::Usage);
+        HelpLevel::Usage.display();
         std::process::exit(0);
     } else {
         parse_cli_from_iter(env::args().skip(1))
@@ -148,62 +240,30 @@ fn parse_cli() -> Result<Config, String> {
 
 fn run() -> Result<(), String> {
     let cli = parse_cli()?;
-    let chunk: usize = (cli.chunk_size as usize) * (cli.sector_size as usize);
-    let buffer: usize = (cli.buffer_size as usize) * chunk;
-
-    // Program welcome - a header for CSV log in case of STDOUT redirection
-    println!("^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^");
-    println!("rusdd - Really Useful Secure Digital Duplicator (ver.0.0.3)");
-    println!("Output Path: {}", cli.output_path.to_string_lossy());
-    if cli.output_path.exists() && !cli.force {
-        return Err(format!(
-            "Output Path ({}) does already exist. \
-            Be careful to check it and back-up if needed. \
-            Then use --force flag if you sure to overwrite it.",
-            cli.output_path.display()
-        ));
-    }
-    if cli.smart {
-        println!(
-            "Smart Truncation Stop: ENABLED (after {} empty sectors)",
-            cli.chunk_size
-        );
-        println!("  Detects both 0x00 and 0xFF empty patterns");
-    } else {
-        println!("Smart Truncation Stop: DISABLED (full copy)");
-    }
-    println!("Physical Drive: {}", cli.physical_drive);
-    println!("Sector size: {} bytes", cli.sector_size);
-    println!("Buffer size: {} MiB", buffer / 1024 / 1024);
-    println!("************************************************************");
+    cli.display();
 
     // Actual logic of the imaging physical device into a file bit-by-bit
-    let mut source = File::open(&cli.physical_drive)
-        .map_err(|e| format!("Cannot open {}: {}", cli.physical_drive, e))?;
-    let mut destination = OpenOptions::new()
-        .write(true)
-        .create(true)
-        .truncate(true)
-        .open(&cli.output_path)
-        .map_err(|e| format!("Failed {}: {}", cli.output_path.display(), e))?;
-    let mut buf = vec![0u8; buffer];
+    let mut source = cli.source.open_for_read()?;
+    let mut destination = cli.destination.open_for_write()?;
+    let (_, _, buffer_size) = cli.calc_sizes();
+    let mut buffer = vec![0u8; buffer_size];
 
     // Starting small - read and write only one buffer
     let bytes_read = source
-        .read(&mut buf)
+        .read(&mut buffer)
         .map_err(|e| format!("Read error: {}", e))?;
     if bytes_read == 0 {
         return Err("Source device is empty".to_string());
     }
     eprintln!("Read {} bytes", bytes_read);
     destination
-        .write_all(&buf[..bytes_read])
+        .write_all(&buffer[..bytes_read])
         .map_err(|e| format!("Write error: {}", e))?;
-    println!("************************************************************");
+    println!("***********************************************************");
     println!(
         "Successfully wrote {} bytes to {}",
         bytes_read,
-        cli.output_path.display()
+        cli.destination.display()
     );
 
     Ok(())